Cleanup of Teams files and update BuildLog
Minor restructure to add an additional module for handling msid to
resolve circular reference in app.js of previous commits. This will also
make it easier to convert to a naked msal-node or passport
custom authentication and authorisation implementation. The new BuildLog
contains some more To-Dos to reflect the end-goal of the Teams/Graph
interface.